Changes to: [CPU] [MMU [Utils] - Added StringUtils class - Implemented instructions:   · LD SP, d16   · LD A, A   · LD (a16), A   · LD A, (a16)

+++ src/main/java/com/azurelithium/gueimboi/memory/RAM.java
@@ -13,7 +13,7 @@ class RAM {
         return ram[address];
     }
 
-    void writeByte(int address, byte value) {
+    void writeByte(int address, int value) {
         ram[address] = value;
     }
